requirements.txt
accelerate==1.12.0
annotated-types==0.7.0
anyio==4.12.0
beautifulsoup4==4.14.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
diffusers==0.36.0
exceptiongroup==1.3.1
filelock==3.20.0
fsspec==2025.12.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.7.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
packaging==25.0
peft==0.18.0
pillow==12.0.0
platformdirs==4.5.1
protobuf==6.33.2
psutil==7.1.3
pydantic==2.12.5
pydantic_core==2.41.5
PySocks==1.7.1
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
sentry-sdk==2.47.0
smmap==5.0.2
soupsieve==2.8
sympy==1.14.0
tokenizers==0.22.1
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.3
triton==3.5.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
wandb==0.23.1
zipp==3.23.0
